backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.55.3
python-multipart==0.0.20
Pillow==11.0.0
pydantic==2.10.3

